Update validate parameter methods into module from app controller
